src/pybot/requirements.txt
annotated-types==0.7.0
APScheduler==3.6.3
CacheControl==0.14.2
cachetools==4.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
firebase-admin==6.5.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-telegram-bot==13.15
pytz==2025.1
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
tornado==6.1
typing_extensions==4.12.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
